requirements.txt
beautifulsoup4==4.10.0
Django==3.2.10
mysql-connector-python==8.0.28
requests==2.26.0
